User creation services

+++ client/internal/pso/player/player.go
@@ -33,14 +33,13 @@ type BasePlayerInfo struct {
 	Warping             bool
 }
 
-func ParsePlayerMemory(buf []uint16, base int) BasePlayerInfo {
+func ParsePlayerMemory(buf []uint16, base uintptr) BasePlayerInfo {
 	shiftaMultiplier := numbers.Float32FromU16(buf[(0x278-base)/2], buf[(0x27A-base)/2])
 	debandMultiplier := numbers.Float32FromU16(buf[(0x278+12-base)/2], buf[(0x278+14-base)/2])
 	class := (buf[(0x961-base)/2] & 0xF00) >> 8
 
 	stateBitfield := buf[(0x33E-base)/2]
 	playerWarping := stateBitfield&0x04 > 0
-	// log.Printf("stateBitfield: 0x%08x, 0x%08x", stateBitfield, buf[(0x33E-base)/2+1])
 
 	return BasePlayerInfo{
 		Room:                buf[(0x028-base)/2],
@@ -63,18 +62,13 @@ func ParsePlayerMemory(buf []uint16, base int) BasePlayerInfo {
 	}
 }
 
-func GetPlayerData(handle w32.HANDLE, playerAddress int) (BasePlayerInfo, error) {
-	base := 0x028
-	max := 0xE4E
-	buf, _, ok := w32.ReadProcessMemory(handle, uintptr(playerAddress+base), uintptr((max-base)+4))
+func GetPlayerData(handle w32.HANDLE, playerAddress uintptr) (BasePlayerInfo, error) {
+	base := uintptr(0x028)
+	max := uintptr(0xE4E)
+	buf, _, ok := w32.ReadProcessMemory(handle, playerAddress+base, (max-base)+4)
 	if !ok {
 		return BasePlayerInfo{}, errors.New("unable to getPlayerData")
 	}
-	// log.Printf("%v/%v", read, uintptr(10+(0x3f0-base)/2))
-	// log.Printf("------------")
-	// for i := 0; i <= len(buf)-8; i += 8 {
-	// 	log.Printf("0x%x, 0x%x, 0x%x, 0x%x, 0x%x, 0x%x, 0x%x, 0x%x", buf[i], buf[i+1], buf[i+2], buf[i+3], buf[i+4], buf[i+5], buf[i+6], buf[i+7])
-	// }
 	basePlayerInfo := ParsePlayerMemory(buf, base)
 	name, err := getCharacterName(handle, playerAddress)
 	if err != nil {
@@ -88,13 +82,10 @@ func GetPlayerData(handle w32.HANDLE, playerAddress int) (BasePlayerInfo, error)
 	}
 	basePlayerInfo.GuildCard = guildcard
 	return basePlayerInfo, nil
-
-	// log.Printf("%v - %v/%v - %v/%v", len(buf), pso.CurrentPlayerData.HP, pso.CurrentPlayerData.MaxHP,
-	// 	pso.CurrentPlayerData.TP, pso.CurrentPlayerData.MaxTP)
 }
 
-func getCharacterName(handle w32.HANDLE, playerAddress int) (string, error) {
-	buf, _, ok := w32.ReadProcessMemory(handle, uintptr(playerAddress+0x428), 24)
+func getCharacterName(handle w32.HANDLE, playerAddress uintptr) (string, error) {
+	buf, _, ok := w32.ReadProcessMemory(handle, playerAddress+0x428, 24)
 	if !ok {
 		return "", errors.New("unable to getCharacterName")
 	}
@@ -112,8 +103,8 @@ func getCharacterName(handle w32.HANDLE, playerAddress int) (string, error) {
 	return name, nil
 }
 
-func getGuildCard(handle w32.HANDLE, playerAddress int) (string, error) {
-	guildCard, err := numbers.ReadString(handle, uintptr(playerAddress+0x92c), 7)
+func getGuildCard(handle w32.HANDLE, playerAddress uintptr) (string, error) {
+	guildCard, err := numbers.ReadString(handle, playerAddress+0x92c, 7)
 	if err != nil {
 		return "", err
 	}
